Add files via upload
This code is a Python program that visualizes different sorting algorithms using the Tkinter library for graphical user interfaces (GUIs). The user can input an array of integers or generate a random array, and then select one of the sorting algorithms to visualize how the data is sorted step by step.